feat: 마이페이지 에러바운더리 적용
