+++ frontend/src/components/MyPage/GuestProfile.tsx
@@ -1,0 +1,10 @@
+const GuestProfile = () => {
+  return (
+    <>
+      <div>게스트님</div>
+      <div>오늘도 방끗과 함께 방긋 웃어요! :)</div>
+    </>
+  );
+};
+
+export default GuestProfile;
